Cleanup: move various hook helper functions to a new "cytoscapeData" utility file
