+++ Site/webapp/wdkCustomization/js/client/hooks/clusterGraph.ts
@@ -6,18 +6,20 @@ import {
   NodeDefinition,
   Stylesheet
 } from 'cytoscape';
-import { noop, orderBy, range } from 'lodash';
+import { noop, orderBy } from 'lodash';
 
 import {
-  EdgeType,
   NodeDisplayType,
   ProteinType,
-  corePeripheralLegendColors,
-  edgeTypeDisplayNames
+  corePeripheralLegendColors
 } from '../utils/clusterGraph';
 import {
+  makePieStyles,
+  makeEdgeData,
+  nodeEntryToCytoscapeData
+} from '../utils/cytoscapeData';
+import {
   EcNumberEntry,
-  EdgeEntry,
   GroupLayout,
   NodeEntry,
   PfamDomainEntry
@@ -29,27 +31,6 @@ const MAX_PIE_SLICES = 16;
 interface CyEffectCallback {
   (cy: Core): (void | (() => void | undefined));
 };
-
-interface NodeData {
-  id: string;
-  corePeripheralColor: string;
-  groupColor: string;
-  speciesColor: string;
-  ecPieColors: string[];
-  ecPieSliceSize: string;
-  pfamPieColors: string[];
-  pfamPieSliceSize: string;
-}
-
-interface EdgeData {
-  id: string;
-  source: string;
-  target: string;
-  type: EdgeType;
-  label: string;
-  eValue: number;
-  score: number;
-}
 
 export function useCytoscapeConfig(
   layout: GroupLayout,
@@ -178,25 +159,6 @@ function useNodes(
   );
 }
 
-function nodeEntryToCytoscapeData(
-  nodeEntry: NodeEntry,
-  layout: GroupLayout,
-  corePeripheralMap: Record<string, ProteinType>,
-  taxonUiMetadata: TaxonUiMetadata,
-  orderedEcNumbers: EcNumberEntry[],
-  ecNumberNPieSlices: number,
-  orderedPfamDomains: PfamDomainEntry[],
-  pfamDomainNPieSlices: number
-): NodeData {
-  return {
-    id: nodeEntry.id,
-    corePeripheralColor: nodeEntryToCorePeripheralColor(nodeEntry, layout, corePeripheralMap),
-    ...nodeEntryToTaxonColors(nodeEntry, layout, taxonUiMetadata),
-    ...nodeEntryToEcNumberPieData(nodeEntry, layout, orderedEcNumbers, ecNumberNPieSlices),
-    ...nodeEntryToPfamDomainPieData(nodeEntry, layout, orderedPfamDomains, pfamDomainNPieSlices),
-  };
-}
-
 function nodeEntryToCorePeripheralColor(
   nodeEntry: NodeEntry,
   { group: { genes } }: GroupLayout,
@@ -206,60 +168,6 @@ function nodeEntryToCorePeripheralColor(
   const proteinType = corePeripheralMap[taxonAbbrev];
 
   return corePeripheralLegendColors[proteinType];
-}
-
-function nodeEntryToTaxonColors(
-  nodeEntry: NodeEntry,
-  { group: { genes } }: GroupLayout,
-  { species }: TaxonUiMetadata
-) {
-  const taxonAbbrev = genes[nodeEntry.id].taxon.abbrev;
-  const nodeSpecies = species[taxonAbbrev];
-
-  return {
-    groupColor: nodeSpecies.groupColor,
-    speciesColor: nodeSpecies.color
-  };
-}
-
-function nodeEntryToEcNumberPieData(
-  nodeEntry: NodeEntry,
-  { group: { genes } }: GroupLayout,
-  orderedEcNumbers: EcNumberEntry[],
-  ecNumberNPieSlices: number
-) {
-  const nodeEcNumbers = genes[nodeEntry.id].ecNumbers;
-
-  const ecPieColors = orderedEcNumbers.slice(0, ecNumberNPieSlices).map(
-    ecNumber => nodeEcNumbers.includes(ecNumber.code)
-      ? ecNumber.color
-      : 'white'
-  );
-
-  return {
-    ecPieColors,
-    ecPieSliceSize: `${(100 / ecNumberNPieSlices)}%`
-  };
-}
-
-function nodeEntryToPfamDomainPieData(
-  nodeEntry: NodeEntry,
-  { group: { genes } }: GroupLayout,
-  orderedPfamDomains: PfamDomainEntry[],
-  pfamDomainNPieSlices: number
-) {
-  const nodePfamDomains = Object.keys(genes[nodeEntry.id].pfamDomains);
-
-  const pfamPieColors = orderedPfamDomains.slice(0, pfamDomainNPieSlices).map(
-    pfamDomain => nodePfamDomains.includes(pfamDomain.accession)
-      ? pfamDomain.color
-      : 'white'
-  );
-
-  return {
-    pfamPieColors,
-    pfamPieSliceSize: `${(100 / pfamDomainNPieSlices)}%`
-  };
 }
 
 function useEdges(layout: GroupLayout): EdgeDefinition[] {
@@ -276,18 +184,6 @@ function useEdges(layout: GroupLayout): EdgeDefinition[] {
         .sort((e1, e2) => e2.data.score - e1.data.score),
     [ layout.edges ]
   );
-}
-
-function makeEdgeData(edgeId: string, edgeEntry: EdgeEntry): EdgeData {
-  return {
-    id: edgeId,
-    source: edgeEntry.queryId,
-    target: edgeEntry.subjectId,
-    type: edgeEntry.T,
-    label: `${edgeTypeDisplayNames[edgeEntry.T]}, evalue=${edgeEntry.E}`,
-    eValue: Number(edgeEntry.E),
-    score: edgeEntry.score
-  };
 }
 
 function useStylesheet(
@@ -435,20 +331,4 @@ function useStylesheet(
     ],
     []
   );
-}
-
-function makePieStyles(nPieSlices: number, dataPrefix: string) {
-  const sliceStyles = range(0, nPieSlices).reduce(
-    (memo, i) => ({
-      ...memo,
-      [`pie-${i + 1}-background-color`]: `data(${dataPrefix}PieColors.${i})`,
-      [`pie-${i + 1}-background-size`]: `data(${dataPrefix}PieSliceSize)`
-    }),
-    {}
-  );
-
-  return {
-    'pie-size': '100%',
-    ...sliceStyles
-  };
 }
